==> 1_Data.Rmd <==
---
title: "Data cleaning"
author: "Jason_Rosado"
date: "`r Sys.Date()`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide 
    code_download: true
    fig_width: 8
    fig_height: 6
    theme: cosmo
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, 
                      warning = FALSE, 
                      message = FALSE, 
                      tidy = TRUE,
                      fig.width = 8, 
                      fig.height = 6)
library(tidyverse)
library(here)
library(RColorBrewer) 
library(ggpubr)
library(cowplot)
library(ggh4x)
library(table1)
library(tableone) 
```

## Socio-demographic data
```{r}
#####################
#SOLOMON ISLANDS    #
#####################

### SELECTING SAMPLE HAPLOTYPES WITH >1% FREQUENCY AND PRESENT IN AT LEAST >=2 SAMPLES
Filter_hap_SI <- readRDS(here("raw_data","SI_seq_tbl_End_Filter.rds")) %>%
  filter(Frequency >= 0.01) 

Exclude_singleton_SI<- Filter_hap_SI%>%
   dplyr::group_by(Haplotype) %>%
   dplyr::summarise(n = dplyr::n(), .groups = "drop") %>%
   dplyr::filter(n > 1L) 

Filter_hap_SI <- Filter_hap_SI %>% 
  filter(Haplotype %in% Exclude_singleton_SI$Haplotype)
  
AmpSeq_samples_data_SI <- read.delim(here("raw_data","sample_table_SI_301220.txt"), header = TRUE, sep = "\t")

Epi_data_SI<- read.csv(here("raw_data","SI_Master_epi_data.csv"), header = T) 

# Merging the AmpSeq run information (copies/uL and Cq) with SI Epi data 

hap_SI_epi<- 
  Filter_hap_SI %>%
  left_join(AmpSeq_samples_data_SI, by = c("sample" = "SampleID"))%>%
  mutate(SampleName = as.integer(SampleName),
         infoD = as.integer(infoD)) %>% 
  left_join(Epi_data_SI, by = c("SampleName" = "studyid", "infoD" = "vis_visit_day"))

#########################
# PERUVIAN COHORT       #
#########################
### SELECTING SAMPLE HAPLOTYPES WITH >1% FREQUENCY AND PRESENT IN AT LEAST >=2 SAMPLES
Filter_hap_PE <- readRDS(here("raw_data", "PE_seq_flt_tbl.rds")) %>%
  filter(frequency >= 0.01)
  
Exclude_singleton_PE<- Filter_hap_PE%>%
   dplyr::group_by(haplotype) %>%
   dplyr::summarise(n = dplyr::n(), .groups = "drop") %>%
   dplyr::filter(n > 1L) 

Filter_hap_PE <- Filter_hap_PE %>% 
  filter(haplotype %in% Exclude_singleton_PE$haplotype)

AmpSeq_samples_data_PE <- read.csv(here("raw_data","PE_Sample_sheet.csv"), header = TRUE)

Epi_data_PE<- read.csv(here("raw_data", "PE_23_05_18_Clinic_data-3y.csv"), header = T) 

#Merging the AmpSeq run information (copies/uL and Cq) with PE Epi data
hap_PE_epi<-
Filter_hap_PE %>%
  left_join(AmpSeq_samples_data_PE, by = c("sample_id" = "SampleID")) %>%
    left_join(Epi_data_PE, by = c("sample_id" = "cod_samp", "PatientName" = "cod_per"))


```

## Cleaning low coverage marker. Eliminating chr04 due to low coverage
```{r, results=FALSE}
#Chr04 was run only in SI but not in Peru

hap_SI_epi <- hap_SI_epi %>%
  filter(!marker_id %in% c("Chr04")) 

```

## Cleaning low covered samples. Keeping samples with more than 60% of amplification success (e.g. >6 markers)
```{r, echo=FALSE}

#Percentage of successful amplification in baseline samples
amp_success_samp_SI<- hap_SI_epi %>% 
    group_by(marker_id, sample) %>% 
    count() %>% 
    group_by(sample) %>%
    summarise(n=n())%>% mutate(amp_success = n/max(n))%>%
  filter(amp_success > 0.6)

sum(amp_success_samp_SI$amp_success < "0.60")

hap_SI_epi <- hap_SI_epi[hap_SI_epi$sample %in% amp_success_samp_SI$sample,]


amp_success_samp_PE<- hap_PE_epi %>%
  group_by(marker_id, sample) %>% 
    count() %>% 
    group_by(sample) %>%
    summarise(n=n())%>% mutate(amp_success = n/max(n)) %>%
  filter(amp_success > 0.6)

sum(amp_success_samp_PE$amp_success < "0.6")

hap_PE_epi <- hap_PE_epi[hap_PE_epi$sample %in% amp_success_samp_PE$sample,]

```

##Description of SI samples. Selecting SI samples with follow up (paired samples)
```{r, echo=FALSE, message=FALSE}
#Selection patients with baseline samples
baseline_patients <- Filter_hap_SI %>%
    filter(infoD==0) %>% 
    select(info) %>% distinct() 

#Selecting patients with follow up samples.     
follow_patients <- Filter_hap_SI %>% 
  filter(!infoD ==0)  %>% 
    select(info, sample) %>% distinct() %>%
    inner_join(baseline_patients, by = c("info")) 

#Merging baseline and follow up samples. There are some patients whose follow up samples were successfully amplified but not the baseline one. We are removing those samples with follow up only because we want to identify recurrent infections.
hap_SI_epi <- hap_SI_epi %>% 
    mutate(follow = ifelse(infoD ==0, "baseline", "follow_up"),
    copies_ul = as.numeric(copies_ul),
    parasites_ul = copies_ul/3,
    pv_clin = as.factor(pv_clin))%>%
    inner_join(baseline_patients, by = c("info")) %>%
    left_join(follow_patients, by = c("info", "sample")) %>%      
    drop_na(trt_label) %>%
    distinct() 

#Dataframe with number of episodes sequenced per patient
patients_episodes_SI<- hap_SI_epi%>%
  group_by(info) %>%
        distinct(sample_id) %>%
  count() %>%
  ungroup() %>% 
  mutate(episodes = n) 

#Selecting some variables to generate a table of patient characteristics
patients_sum <- hap_SI_epi %>% 
  select(info, age_y,gender,trt_label)%>%   
  group_by(info) %>% 
  distinct()

#Selecting some variables to generate a table of samples characteristics
samples_sum <- hap_SI_epi %>% 
  select(info,follow, parasites_ul,pv_clin)%>%   
  group_by(info) %>% 
  distinct()

table1(~ age_y +gender+ trt_label, data = patients_sum, render.continuous = c("Median [Q1, Q3]"))

table_samples_summ<- CreateTableOne(vars = c("parasites_ul","pv_clin"), strata = "follow", data = samples_sum, 
               factorVars =c("pv_clin"))

print(table_samples_summ,showAllLevels = TRUE,
      nonnormal = "parasites_ul")
```

##Description of PE samples. Selecting PE samples with follow up(paired samples)
```{r, echo=FALSE, message=FALSE}

#Dataframe with number of episodes sequenced per patient
patients_episodes<- hap_PE_epi%>%
  select(PatientName,sample_id,com)%>%
   group_by(PatientName,com) %>%
        distinct(sample_id, com) %>%
  count() %>%
  ungroup() %>% 
  mutate(com = ifelse(com ==1, "Cahuide", "Lupuna"),
        episodes = n,
        plus_2epis =ifelse(n ==1, "1", ">1")) 

patients_two_episodes_PE<- hap_PE_epi%>%
  select(PatientName,sample_id,com)%>%
   group_by(PatientName,com) %>%
        distinct(sample_id, com) %>%
  count() %>%
  ungroup() %>% 
  mutate(episodes = n)%>% 
  filter(episodes>1)
  
#Selecting some variables to generate a table of samples characteristics
samples_sum_PE<- hap_PE_epi %>% 
  select(PatientName,sample, date,parasites_uL,clinic_Pvinf,medicamento2)%>%
  group_by(PatientName)%>%
  distinct()%>%
  mutate(date = as.Date(date, format = "%m/%d/%Y")) %>%
  mutate(order =rank(date)) %>%
   mutate(day = factor(ifelse(order== 1,"day_0","day_1"), 
                        levels = c("day_0","day_1"), 
                      labels = c("Day 0","Day X")),
          medicamento2 = case_when(medicamento2 == "C" ~ "CQ+PQ",
                                  medicamento2 == "-1" ~ "No treatment"))%>%
  select(sample,day,
   parasites_uL, clinic_Pvinf,medicamento2) %>% 
  distinct()

#Selecting some variables to generate a table of patient characteristics
Patients_sum_PE<- hap_PE_epi %>% 
  select(PatientName,date,age, sex, com,medicamento2, clinic_Pvinf) %>%
  #Selecting the earliest date because age of patient change with time (it is a following study)
  mutate(date = as.Date(date, format = "%m/%d/%Y") )  %>%
  group_by(PatientName)%>% 
  arrange(date) %>% 
  filter(date == min(date)) %>% 
    distinct() %>%
  mutate(sex = ifelse(sex == 1, "male", "female"),
         com = ifelse(com ==1, "Cahuide", "Lupuna"),
         medicamento2 = case_when(medicamento2 == "C" ~ "CQ+PQ",
                                  medicamento2 == "-1" ~ "No treatment"))%>% 
  select(PatientName,age, com,sex,medicamento2) 


table1(~ age +sex+ medicamento2| com, data = Patients_sum_PE, render.continuous = c("Median [Q1, Q3]"))

#Function to transform numeric to categorical variable

rndr <- function(x, ...) {
    if (is.factor(x) || is.character(x)) {
        c(render.default(x, ...), c(`Overall N`=sum(!is.na(x))))
    } else {
        render.default(x, ...)
    }
}

table1(~ episodes + factor(plus_2epis), data =  patients_episodes, 
       render.missing = NULL,
       render.categorical = "FREQ (PCTnoNA%)",
       render = rndr)

table1(~ parasites_uL + factor(clinic_Pvinf)| day, data = samples_sum_PE,render.continuous = c("Median [Q1, Q3]") )

table_summ_PE<- CreateTableOne(vars = c("parasites_uL","clinic_Pvinf"), strata = "day", data = samples_sum_PE, 
               factorVars =c("clinic_Pvinf"))

print(table_summ_PE,showAllLevels = TRUE,
      nonnormal = "parasites_uL")

```
#Success percentage of MARKERS with more than 80% of success
```{r, message=FALSE}
#Percentage of successful amplification in SI samples

success_ratio_SI<- hap_SI_epi %>% 
    group_by(marker_id, sample_id, follow) %>% 
    count() %>% 
    group_by(follow,marker_id) %>%
    summarise(n=n()) %>% mutate(success = n/max(n))

table1(~ success| follow, data = success_ratio_SI, render.continuous = c("Median [Q1, Q3]"))

table_succes_rate_SI<- CreateTableOne(vars = c("success"), strata = "follow", data = success_ratio_SI)

print(table_succes_rate_SI,showAllLevels = TRUE,
      nonnormal = "success")

#Percentage of successful amplification in PE samples

success_ratio_PE<- hap_PE_epi %>% 
    inner_join(samples_sum_PE, by = c("sample")) %>%     
    group_by(day, marker_id, sample_id) %>% 
    count() %>% 
    group_by(day, marker_id) %>%
    summarise(n=n())%>% mutate(success = n/max(n))

table1(~ success| day, data = success_ratio_PE, render.continuous = c("Median [Q1, Q3]"))

table_succes_rate_PE<- CreateTableOne(vars = c("success"), strata = "day", data = success_ratio_PE)

print(table_succes_rate_PE,showAllLevels = TRUE,
      nonnormal = "success")

```
```{r echo=F, eval=F}
dir.create("intermediate_data")
```

```{r}

write_rds(amp_success_samp_SI,"intermediate_data/sols_amp_success.rds", "xz")
write_rds(amp_success_samp_PE,"intermediate_data/peru_amp_success.rds", "xz")
```

```{r}
save.image("./intermediate_data/intermediate_data.RData")
```
