==> 15-Pv3Rs_additional_sensitivity_analysis.Rmd <==
---
title: "Pv3Rs additional sensitivity analyses"
author: "Shazia Ruybal-Pesántez"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(Pv3Rs)
library(patchwork)

source("helper_functions.R")
```

# Solomon Islands
```{r data, eval=F}
# load("outputs/Pv3Rs_sols_posteriors_20240827_181146.RData")
# save(analysis_data, baseline_fs, file = "outputs/sols_sensitivity_input.RData")

load("outputs/sols_sensitivity_input.Rdata")
```

## Prior probabilities
Aim: to see how well the model works by skewing prior probabilities completely to one or the other (recrudescence vs relapse) to see impact on results
```{r compute probabilities}
# Skewing completely to one probable classification scenario
priors <- list(
  relapse_only       = c(C = 0, L = 1, I = 0),
  recrudescence_only = c(C = 1, L = 0, I = 0),
  reinfection_only   = c(C = 0, L = 0, I = 1)
)

# compute Pv3Rs for each scenario
results <- imap(priors, function(prior_vec, scenario) {
  # build a fresh global_vars list for this prior
  gv <- set_global_vars(PRIOR_3RS = prior_vec)
  
  compute_all(
    analysis_data = analysis_data,
    global_vars   = gv,
    fs            = baseline_fs,
    filename      = paste0("./outputs/Pv3Rs_sols_posteriors_", scenario, "_")
  )
})

# get marginal summaries for each prior type in one df
marginal_summary <- imap_dfr(results, ~ .x %>%
                              pluck("indiv_posteriors_marginal") %>%
                              get_marginal_summary() %>%
                              mutate(prior_type = .y)
)

# now joint summaries
joint_summary    <- imap_dfr(results, ~ .x %>%
                              pluck("indiv_posteriors_joint") %>%
                              get_joint_summary() %>%
                              mutate(prior_type = .y)
)
```

```{r results}
marginal_summary %>% 
  tabyl(prior_type, posterior_value) %>% 
  adorn_totals("col")
```
Not sure I understand why there are NaNs.

```{r}
plot_episode_prob(marginal_summary) + facet_wrap(subject_id ~ prior_type)
```


## Marker-by-marker removal
### Run
Aim: to see if there is a marker-specific impact on the probability of recrudescence vs relapse classification
```{r prep and compute, eval=F}
# Removing marker one at a time

panel <- c("Chr05","Chr07","Chr09","Chr10","Chr08", "Chr13","Chr11","Chr03","Chr01","Chr02","Chr14")

# compute Pv3Rs after removing one marker 
results_rm_mhap <- map(panel, function(m) {
  
  # panel without the marker
  bm <- setdiff(panel, m)
  
  # rebuild global vars
  gv <- set_global_vars(BENCHMARK_MARKERS = bm)
  
  compute_all(
    analysis_data = analysis_data,
    fs            = baseline_fs,
    global_vars   = gv,
    filename      = paste0("./outputs/Pv3Rs_sols_posteriors_remove", m, "_")
  )
}) %>% 
  set_names(panel)

# get marginal summaries for removed mhap in one df
marginal_summary_rm_mhap <- imap_dfr(results_rm_mhap, ~ .x %>% 
                                       pluck("indiv_posteriors_marginal") %>% 
                                       get_marginal_summary() %>%
                                       mutate(removed_mhap = .y)
)

# same for joint
joint_summary_rm_mhap <- imap_dfr(results_rm_mhap, ~ .x %>% 
                                       pluck("indiv_posteriors_joint") %>% 
                                       get_joint_summary() %>%
                                       mutate(removed_mhap = .y)
)
```

```{r save, eval=F}
save(analysis_data, results_rm_mhap, marginal_summary_rm_mhap, joint_summary_rm_mhap,
     file = "./outputs/sols_mhap_sensitivity_outputs.RData")
```

### Results
```{r load}
load("outputs/sols_mhap_sensitivity_outputs.RData")
load("outputs/sols_infection_classification.RData")
```

```{r}
# order mhaps by most He to least
panel_he <- c("Chr05", "Chr02", "Chr07", "Chr14", "Chr13", "Chr09", "Chr10", "Chr03", "Chr11", "Chr01", "Chr08")
  
plot_mhap_sens_sols <- marginal_summary_rm_mhap %>%
  mutate(
    removed_mhap = factor(removed_mhap, levels = panel_he),
    episode_number = factor(episode_number)  
  ) %>% 

ggplot(aes(x = removed_mhap,
           y = posterior_value,
           fill = posterior_classification)) +
  geom_bar(stat = "identity", position = "fill") +
  scale_fill_manual(values = c("Relapse"       = "turquoise3",
                               "Recrudescence" = "skyblue4",
                               "Reinfection"   = "magenta3")) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  labs(x    = "Removed microhaplotype (ordered from highest to lowest He)",
    y    = "Posterior probability",
    fill = "") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +

  facet_wrap(subject_id ~ episode_number)      

ggsave("outputs/plots/supp_marker_sensitivity_sols.tiff", plot_mhap_sens_sols, width = 13, height = 9)
```

```{r possible misclassification}
misclassified_ids <- tibble(subject_id       = character(),
                            episode_number   = integer(),
                            problematic_mhap = character()) %>% 
  add_row(subject_id = "AR-098",
          episode_number = 2,
          problematic_mhap = "Chr10") %>% 
  add_row(subject_id = "AR-133",
          episode_number = 3,
          problematic_mhap = "Chr07") %>% 
  add_row(subject_id = "AR-140",
          episode_number = 2,
          problematic_mhap = "Chr05") %>% 
  add_row(subject_id = "AR-142",
          episode_number = 3,
          problematic_mhap = "Chr07") %>% 
  add_row(subject_id = "AR-173",
          episode_number = 2,
          problematic_mhap = "Chr07") %>% 
  add_row(subject_id = "AR-176",
          episode_number = 3,
          problematic_mhap = "Chr09") %>% 
  add_row(subject_id = "AR-214",
          episode_number = 2,
          problematic_mhap = "Chr10") %>% 
  add_row(subject_id = "AR-258",
          episode_number = 2,
          problematic_mhap = "Chr10") %>% 
  add_row(subject_id = "AR-302",
          episode_number = 2,
          problematic_mhap = "Chr05") %>% 
  add_row(subject_id = "AR-306",
          episode_number = 2,
          problematic_mhap = "Chr10") %>% 
  add_row(subject_id = "AR-325",
          episode_number = 2,
          problematic_mhap = "Chr09") %>% 
  add_row(subject_id = "AR-326",
          episode_number = 2,
          problematic_mhap = "Chr10") 

```

```{r}
# 1) build one row per subject/marker with the baseline haps
baseline_haps <- analysis_data %>%
  filter(str_ends(sample_id, "-0")) %>%
  group_by(subject_id, marker_id) %>%
  summarize(baseline_haps = list(unique(haplotype)), .groups="drop")

# 2) join & flag per-haplotype diffs
per_hap <- analysis_data %>%
  left_join(baseline_haps, by = c("subject_id","marker_id")) %>%
  left_join(classification_summary_sols %>% 
              select(sampleid1, joint_classification),
            by = join_by(sample_id == sampleid1)) %>% 
  rowwise() %>%
  mutate(diff_from_baseline = !(haplotype %in% baseline_haps)) %>%
  ungroup() 

# 3) collapse to one row per sample–marker: did *any* hap differ?
plot_df <- per_hap %>%
  group_by(subject_id, sample_id, episode_number, marker_id, joint_classification) %>%
  summarize(any_diff = any(diff_from_baseline),
    .groups = "drop") %>%
  # define a fill variable: color all diffs, grey all non-diffs
  mutate(
    fill_flag = ifelse(any_diff, "different", "same")
  ) %>% 
  left_join(misclassified_ids, by = join_by(subject_id, episode_number))

# 4) plot
plot_mhap_trend_sols <-plot_df %>% 
  # focus only on 1st recurrence and remove NAs (eg 359 that had MOI>8)
  filter(episode_number %in% c(1, 2)) %>% 
  filter(!subject_id == "AR-359") %>% 
  # 'fill in' classification for baseline for nice plotting
  group_by(subject_id) %>% 
  fill(joint_classification, .direction = "downup") %>% 
  mutate(marker_id = factor(marker_id, levels = panel_he)) %>% 
  ggplot(aes(x = marker_id, y = reorder(factor(sample_id), episode_number))) +
    geom_tile(aes(fill = fill_flag), colour = "black") +
    scale_fill_manual(values = c(
      same    = "grey80",
      different  = "firebrick"
    )) +
    geom_point(data = plot_df %>% 
                 # focus only on 1st recurrence and remove NAs (eg 359 that had MOI>8)
                 filter(episode_number %in% c(1, 2)) %>% 
                 filter(!subject_id == "AR-359") %>% 
                 # 'fill in' classification for baseline for nice plotting
                 group_by(subject_id) %>% 
                 fill(joint_classification, .direction = "downup") %>% 
                 filter(!is.na(problematic_mhap)) %>% 
                 distinct(subject_id, sample_id, episode_number, joint_classification, marker_id, problematic_mhap) %>% 
                 filter(marker_id == problematic_mhap),
              color = "white",
              shape = 8) +
    facet_wrap(subject_id ~ joint_classification, scales = "free_y", ncol = 4) +
    theme_bw() +
    theme(
      axis.text.x = element_text(angle = 90, hjust = 1),
      strip.background = element_blank()
    ) +
    labs(    
      x    = "Marker (ordered from highest to lowest He)",
      y    = "Sample ID",
      fill = "Compared to baseline haplotype(s)",
      #subtitle = "Note: this plot simplifies the overall trends (e.g. if any haplotype is different to baseline it is colored in red, but we do not plot the haplotype-specific information where more than one haplotype may be possible).",
      caption = "Points with * denote markers that when removed during sensitivity analyses impacted the classification results"
    )

ggsave("outputs/plots/supp_genetic_data_sols.tiff", plot_mhap_trend_sols, width = 13, height = 9)
```

# Peru
```{r data, eval=F}
# load("outputs/Pv3Rs_peru_posteriors_20240904_010047.RData")
# save(analysis_data, fs, file = "outputs/peru_sensitivity_input.RData")

load("outputs/peru_sensitivity_input.Rdata")
```

## Prior probabilities
Aim: to see how well the model works by skewing prior probabilities completely to one or the other (recrudescence vs relapse) to see impact on results
```{r compute probabilities}
# Skewing completely to one probable classification scenario
priors <- list(
  relapse_only       = c(C = 0, L = 1, I = 0),
  recrudescence_only = c(C = 1, L = 0, I = 0),
  reinfection_only   = c(C = 0, L = 0, I = 1)
)

# compute Pv3Rs for each scenario
results <- imap(priors, function(prior_vec, scenario) {
  # build a fresh global_vars list for this prior
  gv <- set_global_vars(PRIOR_3RS = prior_vec)
  
  compute_all(
    analysis_data = analysis_data,
    global_vars   = gv,
    fs            = fs,
    filename      = paste0("./outputs/Pv3Rs_peru_posteriors_", scenario, "_")
  )
})

# get marginal summaries for each prior type in one df
marginal_summary <- imap_dfr(results, ~ .x %>%
                              pluck("indiv_posteriors_marginal") %>%
                              get_marginal_summary() %>%
                              mutate(prior_type = .y)
)

# now joint summaries
joint_summary    <- imap_dfr(results, ~ .x %>%
                              pluck("indiv_posteriors_joint") %>%
                              get_joint_summary() %>%
                              mutate(prior_type = .y)
)
```


```{r results}
marginal_summary %>% 
  tabyl(prior_type, posterior_value) %>% 
  adorn_totals("col")
```
Not sure I understand why there are NaNs.

```{r}
plot_episode_prob(marginal_summary) + facet_wrap(subject_id ~ prior_type)
```

## Marker-by-marker removal
### Run
Aim: to see if there is a marker-specific impact on the probability of recrudescence vs relapse classification
```{r prep and compute, eval=F}
# Removing marker one at a time

panel <- c("Chr05","Chr07","Chr09","Chr10","Chr08", "Chr13","Chr11","Chr03","Chr01","Chr02","Chr14")

# compute Pv3Rs after removing one marker 
results_rm_mhap <- map(panel, function(m) {
  
  # panel without the marker
  bm <- setdiff(panel, m)
  
  # rebuild global vars
  gv <- set_global_vars(BENCHMARK_MARKERS = bm,
                        PRIOR_3RS = c("C" = 0.10, "L" = 0.45, "I" = 0.45))
  
  compute_all(
    analysis_data = analysis_data,
    fs            = fs,
    global_vars   = gv,
    filename      = paste0("./outputs/Pv3Rs_peru_posteriors_remove", m, "_")
  )
}) %>% 
  set_names(panel)

# get marginal summaries for removed mhap in one df
marginal_summary_rm_mhap <- imap_dfr(results_rm_mhap, ~ .x %>% 
                                       pluck("indiv_posteriors_marginal") %>% 
                                       get_marginal_summary() %>%
                                       mutate(removed_mhap = .y)
)

# same for joint
joint_summary_rm_mhap <- imap_dfr(results_rm_mhap, ~ .x %>% 
                                       pluck("indiv_posteriors_joint") %>% 
                                       get_joint_summary() %>%
                                       mutate(removed_mhap = .y)
)
```

```{r save, eval=F}
save(analysis_data, results_rm_mhap, marginal_summary_rm_mhap, joint_summary_rm_mhap,
     file = "./outputs/peru_mhap_sensitivity_outputs.RData")
```

### Results
```{r load}
load("outputs/peru_mhap_sensitivity_outputs.RData")
load("outputs/peru_infection_classification.RData")
```

```{r}
# order mhaps by most He to least
panel_he <- c("Chr05", "Chr11", "Chr07", "Chr08", "Chr13", "Chr02",  "Chr14",  "Chr09", "Chr03", "Chr01", "Chr10")
  
plot_mhap_sens_peru <- marginal_summary_rm_mhap %>%
  mutate(
    removed_mhap = factor(removed_mhap, levels = panel_he),
    episode_number = factor(episode_number)  # treat episode as discrete
  ) %>% 

ggplot(aes(x = removed_mhap,
           y = posterior_value,
           fill = posterior_classification)) +
  geom_bar(stat = "identity", position = "fill") +
  scale_fill_manual(values = c("Relapse"       = "turquoise3",
                               "Recrudescence" = "skyblue4",
                               "Reinfection"   = "magenta3")) +
  scale_x_discrete(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  labs(x    = "Removed microhaplotype (ordered from highest to lowest He)",
       y    = "Posterior probability",
       fill = "") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  facet_wrap(subject_id ~ episode_number)      

ggsave("outputs/plots/supp_marker_sensitivity_peru.tiff", plot_mhap_sens_peru, width = 13, height = 9)
```

```{r possible misclassification}
misclassified_ids <- tibble(subject_id       = character(),
                            episode_number   = integer(),
                            problematic_mhap = character()) %>% 
  add_row(subject_id = "M1B011B",
          episode_number = 2,
          problematic_mhap = "Chr07") %>% 
  add_row(subject_id = "M1C081A",
          episode_number = 3,
          problematic_mhap = "Chr01") %>% 
  add_row(subject_id = "M1C090C",
          episode_number = 2,
          problematic_mhap = "Chr05") 
```

```{r}
# 1) build one row per subject/marker with all haps
all_haps <- analysis_data %>%
  filter(timepoint=="Day 0") %>%
  group_by(subject_id, marker_id) %>%
  summarize(all_haps = list(unique(haplotype)), .groups="drop")

# 2) join & flag per-haplotype diffs
per_hap <- analysis_data %>%
  left_join(all_haps, by = c("subject_id","marker_id")) %>%
  left_join(classification_summary_peru %>% 
              select(sampleid1, joint_classification),
            by = join_by(sample_id == sampleid1)) %>% 
  rowwise() %>%
  mutate(diff_from_baseline = !(haplotype %in% all_haps)) %>%
  ungroup()

# 3) collapse to one row per sample–marker: did *any* hap differ?
plot_df <- per_hap %>%
  group_by(subject_id, sample_id, episode_number, marker_id, joint_classification) %>%
  summarize(any_diff = any(diff_from_baseline),
            .groups = "drop") %>%
  # define a fill variable: color all diffs, grey all non-diffs
  mutate(fill_flag = ifelse(any_diff, "different", "same")) %>% 
  left_join(misclassified_ids, by = join_by(subject_id, episode_number))

# 4) plot
plot_mhap_trend_peru <- plot_df %>% 
  # focus only on 1st recurrence
  filter(episode_number %in% c(1,2)) %>% 
  # 'fill in' classification for baseline for nice plotting
  group_by(subject_id) %>% 
  fill(joint_classification, .direction = "downup") %>% 
  
  mutate(marker_id = factor(marker_id, levels = panel_he)) %>% 
  ggplot(aes(x = marker_id, y = reorder(factor(sample_id), episode_number))) +
    geom_tile(aes(fill = fill_flag), colour = "black") +
    scale_fill_manual(values = c(same    = "grey80",
                                 different  = "firebrick")) +
    geom_point(data = plot_df %>% 
                 # focus only on 1st recurrence
                 filter(episode_number %in% c(1,2)) %>% 
                 # 'fill in' classification for baseline for nice plotting
                 group_by(subject_id) %>% 
                 fill(joint_classification, .direction = "downup") %>% 
                 filter(!is.na(problematic_mhap)) %>% 
                 distinct(subject_id, sample_id, episode_number, joint_classification, marker_id, problematic_mhap) %>% 
                 filter(marker_id == problematic_mhap),
              color = "white",
              shape = 8) +
    facet_wrap(subject_id ~ joint_classification, scales = "free_y", ncol = 4) +
    theme_bw() +
    theme(axis.text.x = element_text(angle = 90, hjust = 1),
      strip.background = element_blank()) +
    labs(    
      x    = "Marker (ordered from highest to lowest He)",
      y    = "Sample ID",
      fill = "Compared to baseline haplotype(s)",
      # subtitle = "Note: this plot simplifies the overall trends (e.g. if any haplotype is different to baseline it is colored in red, but we do not plot the haplotype-specific information where more than one haplotype may be possible).",
      caption = "Points with * denote markers that when removed during sensitivity analyses impacted the classification results"
    )

ggsave("outputs/plots/supp_genetic_data_peru.tiff", plot_mhap_trend_peru, width = 13, height = 12)
```

## Save supp figures
```{r}
plot_mhap_sens_sols / plot_mhap_sens_peru +
  plot_annotation(tag_levels = "A") + plot_layout(guides = "collect")

ggsave("outputs/plots/supp_marker_sensitivity.tiff", width = 12, height = 16)

plot_mhap_trend_sols / plot_mhap_trend_peru +
  plot_annotation(tag_levels = "A")

ggsave("outputs/plots/supp_marker_trends.tiff", width = 12, height = 20)
```
